requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.2.0
httpx==0.28.0
Pillow==11.0.0
numpy==2.2.1
torch==2.5.1
transformers==4.47.1
